app/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-restful
flask-cors
pyjwt
python-dotenv